Refactor blog post preview into own component
